requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
sqlalchemy>=2.0.0
pymysql>=1.0.0
python-jose[cryptography]>=3.3.0
passlib[argon2]>=1.7.4
python-multipart>=0.0.6
python-dotenv>=1.0.0
pypdf>=3.0.0
requests>=2.31.0
pydantic>=2.0.0
email-validator>=2.0.0
pulp>=2.7.0
numpy>=1.20.0
pandas>=2.0.0
scikit-learn>=1.2.0
gensim>=4.3.0
nltk>=3.8.0
networkx>=3.1
